Developer-Fixes/Scripts/requirement.txt
requests==2.24.0
beautifulsoup4==4.9.1
solidity-parser==0.0.7
semantic_version==2.8.5
attrs==20.1.0
slither-analyzer==0.6.12
deepdiff==5.0.2
GitPython==3.1.11
